+++ osu!stream/osu!stream_android.csproj
@@ -77,7 +77,7 @@
     <Compile Include="AssetManager\NativeAssetManagerAndroid.cs" />
     <Compile Include="Audio\BackgroundAudioPlayerAndroid.cs" />
     <Compile Include="Audio\BassNetUtils\Utils.cs" />
-    <Compile Include="Audio\SoundEffectPlayerBass.cs" />
+    <Compile Include="Audio\SoundEffectPlayerBassAndroid.cs" />
     <Compile Include="Audio\SoundEffectPlayerOpenAL.cs" />
     <Compile Include="Graphics\Renderers\NativeTextRendererAndroid.cs" />
     <Compile Include="Input\Sources\InputSourceAndroid.cs" />
